requirements.txt
streamlit
groq
PyPDF2
sqlalchemy
sentence-transformers
faiss-cpu
numpy
